[Canvas V2][5/n] Improve getHoveredEvent clarity
